test: centralize FakeRequest and shared SSE fixtures for ai repository tests
- FakeRequest moved to test/mocks/mocks.dart, removed from 4 inference
  repository tests
- createSseStreamedResponse/createSseChunkEvent/createSseFinalEvent
  deduplicated into test/features/ai/repository/sse_test_utils.dart
