+++ test/features/ai/TEST_REVIEW.md
@@ -61,7 +61,7 @@ Counting methodology: each distinct finding per sub-report is counted once; find
 
 - [x] **[HIGH]** **Eliminate `pumpAndSettle` clusters in `ui/settings/`**: 114 calls in `inference_provider_edit_page_test.dart`, 22 in `provider_filter_chips_row_test.dart`, 20 in `ai_config_delete_service_test.dart`, 16 in `ai_pick_provider_modal_test.dart`, 15 in `modality_selection_modal_test.dart`, 15 in `mlx_audio_model_download_dialog_test.dart`, 9 in both `ai_settings_tab_bar_test.dart` and `ai_settings_floating_action_button_test.dart`. Replace with `tester.pump()` or `tester.pump(kThemeAnimationDuration)` throughout — eliminates 10-second hang risk on every call. **RESOLVED:** every listed file is settle-free or bounded now — inference_provider_edit_page (114→3, the rest genuine route transitions), provider_filter_chips_row/mlx_audio_model_download_dialog/ai_settings_tab_bar/ai_settings_floating_action_button at 0, and this pass converted ai_config_delete_service (38), modality_selection_modal (23), and ai_pick_provider_modal (16) to pump + 300ms bounded pumps.
 
-- [ ] **[HIGH]** **Centralise all inline mock classes** into `test/mocks/mocks.dart`. Confirmed inline duplicates include: `MockOllamaInferenceRepository`, `MockConversationStrategy`, `MockAiConfigDb` (×2 files), `MockTaskProgressRepository` (×3 files), `MockUnifiedAiInferenceRepository` (×2 files), `MockAiConfigByTypeController`, `FakeRequest` (×2 files), and multiple SSE/HTTP fakes. Deduplicate inline mock in the SSE layer: `createSseStreamedResponse`/`createSseChunkEvent` helpers are copy-pasted verbatim between `mistral_inference_repository_test.dart` and `voxtral_inference_repository_test.dart` — extract to shared `test_utils.dart`.
+- [x] **[HIGH]** **Centralise all inline mock classes** into `test/mocks/mocks.dart`. Confirmed inline duplicates include: `MockOllamaInferenceRepository`, `MockConversationStrategy`, `MockAiConfigDb` (×2 files), `MockTaskProgressRepository` (×3 files), `MockUnifiedAiInferenceRepository` (×2 files), `MockAiConfigByTypeController`, `FakeRequest` (×2 files), and multiple SSE/HTTP fakes. Deduplicate inline mock in the SSE layer: `createSseStreamedResponse`/`createSseChunkEvent` helpers are copy-pasted verbatim between `mistral_inference_repository_test.dart` and `voxtral_inference_repository_test.dart` — extract to shared `test_utils.dart`. **RESOLVED:** `MockOllamaInferenceRepository`, `MockConversationStrategy`, `MockAiConfigDb`, `MockTaskProgressRepository`, `MockUnifiedAiInferenceRepository` were already centralised in earlier passes; `MockAiConfigByTypeController` lives in the shared `test/features/ai/test_utils.dart`. This pass: `FakeRequest` moved to `test/mocks/mocks.dart` (removed from cloud/voxtral/mistral/ollama inference repository tests) and the verbatim `createSseStreamedResponse`/`createSseChunkEvent`/`createSseFinalEvent` copies extracted to `test/features/ai/repository/sse_test_utils.dart` (superset signature, shared by mistral + voxtral). Verified no `class Mock*` in any ai test file duplicates a central mock.
 
 - [x] **[HIGH]** **Consolidate five split test files** that violate the one-test-file-per-source rule back into two canonical files: `ai_input_repository_test.dart` (absorb `_language_test`, `_suppressed_ids_test`, and the misplaced `ai_linked_task_context_test.dart` groups) and `ai_config_repository_test.dart` (absorb `_cascade_delete_test` and `_integration_test`). Simultaneously fix the four `Future.delayed(Duration.zero)` violations in these files.
